scripts: Handle non-existent `include` `include_gen` `src_gen` dirs
The user may want to delete them to be sure that everything is
regenerated from scratch.

+++ scripts/collector.py
@@ -193,7 +193,8 @@ def extract_generate_xcbproto():
 def main():
     download()
     print(f'Cleaning up {INCLUDE_PATH}.')
-    shutil.rmtree(INCLUDE_PATH)
+    if path.isdir(INCLUDE_PATH):
+        shutil.rmtree(INCLUDE_PATH)
     extract()
     extract_generate_xcbproto()
